+++ injectable_generator/lib/resolvers/importable_type_resolver.dart
@@ -10,7 +10,8 @@ abstract class ImportableTypeResolver {
 
   ImportableType resolveType(DartType type);
 
-  ImportableType resolveFunctionType(FunctionType function, [ExecutableElement? executableElement]);
+  ImportableType resolveFunctionType(FunctionType function,
+      [ExecutableElement? executableElement]);
 
   static String? relative(String? path, Uri? to) {
     if (path == null || to == null) {
@@ -18,12 +19,16 @@ abstract class ImportableTypeResolver {
     }
     var fileUri = Uri.parse(path);
     var libName = to.pathSegments.first;
-    if ((to.scheme == 'package' && fileUri.scheme == 'package' && fileUri.pathSegments.first == libName) ||
+    if ((to.scheme == 'package' &&
+            fileUri.scheme == 'package' &&
+            fileUri.pathSegments.first == libName) ||
         (to.scheme == 'asset' && fileUri.scheme != 'package')) {
       if (fileUri.path == to.path) {
         return fileUri.pathSegments.last;
       } else {
-        return p.posix.relative(fileUri.path, from: to.path).replaceFirst('../', '');
+        return p.posix
+            .relative(fileUri.path, from: to.path)
+            .replaceFirst('../', '');
       }
     } else {
       return path;
@@ -56,7 +61,8 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
     }
     libs.where((e) => e.exportNamespace.definedNames.values.contains(element));
     for (var lib in libs) {
-      if (!_isCoreDartType(lib) && lib.exportNamespace.definedNames.values.contains(element)) {
+      if (!_isCoreDartType(lib) &&
+          lib.exportNamespace.definedNames.values.contains(element)) {
         imports.add(lib.identifier);
       }
     }
@@ -68,8 +74,10 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
   }
 
   @override
-  ImportableType resolveFunctionType(FunctionType function, [ExecutableElement? executableElement]) {
-    final functionElement = executableElement ?? function.element ?? function.alias?.element;
+  ImportableType resolveFunctionType(FunctionType function,
+      [ExecutableElement? executableElement]) {
+    final functionElement =
+        executableElement ?? function.element ?? function.alias?.element;
     if (functionElement == null) {
       throw 'Can not resolve function type \nTry using an alias e.g typedef MyFunction = ${function.getDisplayString(withNullability: false)};';
     }
@@ -95,15 +103,20 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
   List<ImportableType> _resolveTypeArguments(DartType typeToCheck) {
     final importableTypes = <ImportableType>[];
     if (typeToCheck is RecordType && typeToCheck.alias == null) {
-      for (final recordField in [...typeToCheck.positionalFields, ...typeToCheck.namedFields]) {
+      for (final recordField in [
+        ...typeToCheck.positionalFields,
+        ...typeToCheck.namedFields
+      ]) {
         final imports = resolveImports(recordField.type.element);
         importableTypes.add(ImportableType(
           name: recordField.type.element?.name ?? 'void',
           import: imports.firstOrNull,
           otherImports: imports.skip(1).toSet(),
-          isNullable: recordField.type.nullabilitySuffix == NullabilitySuffix.question,
+          isNullable:
+              recordField.type.nullabilitySuffix == NullabilitySuffix.question,
           typeArguments: _resolveTypeArguments(recordField.type),
-          nameInRecord: recordField is RecordTypeNamedField ? recordField.name : null,
+          nameInRecord:
+              recordField is RecordTypeNamedField ? recordField.name : null,
         ));
       }
     } else if (typeToCheck is ParameterizedType || typeToCheck.alias != null) {
@@ -127,7 +140,8 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
           importableTypes.add(ImportableType(name: 'dynamic'));
         } else {
           importableTypes.add(ImportableType(
-            name: type.element?.name ?? type.getDisplayString(withNullability: false),
+            name: type.element?.name ??
+                type.getDisplayString(withNullability: false),
             import: imports.firstOrNull,
             otherImports: imports.skip(1).toSet(),
             isNullable: type.nullabilitySuffix == NullabilitySuffix.question,
@@ -153,7 +167,8 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
       );
     }
     return ImportableType(
-      name: effectiveElement?.displayName ?? type.getDisplayString(withNullability: false),
+      name: effectiveElement?.displayName ??
+          type.getDisplayString(withNullability: false),
       isNullable: type.nullabilitySuffix == NullabilitySuffix.question,
       import: imports.firstOrNull,
       otherImports: imports.skip(1).toSet(),